requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
psycopg[binary]
PyMuPDF
langchain
langchain-core
langchain-community
langchain-postgres
langchain-huggingface
langchain-ollama
huggingface-hub
sentence-transformers
pydantic
twilio
google-auth
google-auth-oauthlib
python-jose
starlette
requests
